+++ Test/CoreSDK.Test/Shared/Extensibility/TelemetryConfigurationFactoryTest.cs
@@ -18,6 +18,15 @@
     [TestClass]
     public class TelemetryConfigurationFactoryTest
     {
+        private const string EnvironmentVariableName = "APPINSIGHTS_INSTRUMENTATIONKEY";
+
+        [TestCleanup]
+        public void TestCleanup()
+        {
+            Environment.SetEnvironmentVariable(EnvironmentVariableName, null);
+            PlatformSingleton.Current = null; // Force reinitialization in future tests so that new environment variables will be loaded.
+        }
+
         #region Instance
 
         [TestMethod]
@@ -54,7 +63,7 @@
         [TestMethod]
         public void InitializeCreatesInMemoryChannel()
         {
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null);
 
             Assert.IsType<InMemoryChannel>(configuration.TelemetryChannel);
@@ -65,7 +74,7 @@
         {
             string configFileContents = Configuration("<InstrumentationKey>F8474271-D231-45B6-8DD4-D344C309AE69</InstrumentationKey>");
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, new TestableTelemetryModules(), configFileContents);
 
             // Assume that LoadFromXml method is called, tested separately
@@ -76,7 +85,7 @@
         [TestMethod]
         public void InitializeAddsOperationContextTelemetryInitializerByDefault()
         {
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null);
 
             var contextInitializer = configuration.TelemetryInitializers[0];
@@ -99,7 +108,7 @@
         [TestMethod]
         public void InitializeCreatesInMemoryChannelEvenWhenConfigIsBroken()
         {
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null, Configuration("</blah>"));
 
             Assert.IsType<InMemoryChannel>(configuration.TelemetryChannel);
@@ -108,10 +117,58 @@
         [TestMethod]
         public void InitializeCreatesInMemoryChannelEvenWhenConfigIsInvalid()
         {
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null, Configuration("<blah></blah>"));
 
             Assert.IsType<InMemoryChannel>(configuration.TelemetryChannel);
+        }
+
+        [TestMethod]
+        public void InitializeReadsInstrumentationKeyFromEnvironmentVariableIfNotSpecifiedInConfig()
+        {
+            // ARRANGE
+            var ikey = Guid.NewGuid().ToString();
+            Environment.SetEnvironmentVariable(EnvironmentVariableName, ikey);
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
+
+            // ACT
+            new TestableTelemetryConfigurationFactory().Initialize(configuration, null);
+
+            // ASSERT
+            Assert.Equal(ikey, configuration.InstrumentationKey);
+        }
+
+        [TestMethod]
+        public void InitializeDoesNotReadInstrumentationKeyFromEnvironmentVariableIfSpecifiedInConfig()
+        {
+            // ARRANGE
+            var ikeyConfig = Guid.NewGuid().ToString();
+            var ikeyEnvironmentVariable = Guid.NewGuid().ToString();
+
+            Environment.SetEnvironmentVariable(EnvironmentVariableName, ikeyEnvironmentVariable);
+
+            TelemetryConfiguration configuration = new TelemetryConfiguration() { InstrumentationKey = ikeyConfig };
+
+            // ACT
+            new TestableTelemetryConfigurationFactory().Initialize(configuration, null);
+
+            // ASSERT
+            Assert.Equal(ikeyConfig, configuration.InstrumentationKey);
+        }
+
+        [TestMethod]
+        public void InitializeLeavesInstrumentationKeyEmptyWhenNotSpecifiedInConfigOrInEnvironmentVariable()
+        {
+            // ARRANGE
+            Environment.SetEnvironmentVariable(EnvironmentVariableName, null);
+            
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
+
+            // ACT
+            new TestableTelemetryConfigurationFactory().Initialize(configuration, null);
+
+            // ASSERT
+            Assert.Equal(string.Empty, configuration.InstrumentationKey);
         }
 
         #endregion
@@ -141,7 +198,7 @@
         [TestMethod]
         public void CreateInstanceReturnsNullWhenInstanceDoesNotImplementExpectedInterface()
         {
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             Type invalidType = typeof(object);
             Assert.Null(TestableTelemetryConfigurationFactory.CreateInstance(typeof(ITelemetryInitializer), invalidType.AssemblyQualifiedName));
         }
@@ -156,7 +213,7 @@
             string expected = Guid.NewGuid().ToString();
             string profile = Configuration("<InstrumentationKey>" + expected + "</InstrumentationKey>");
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             TestableTelemetryConfigurationFactory.LoadFromXml(configuration, null, XDocument.Parse(profile));
 
             // Assume LoadFromXml calls LoadInstance, which is tested separately.
@@ -330,7 +387,7 @@
                   "<Add Type=\"" + typeof(StubTelemetryProcessor2).AssemblyQualifiedName + "\" />" +
                   "</TelemetryProcessors>");
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, new TestableTelemetryModules(), configFileContents);
 
             // Assume that LoadFromXml method is called, tested separately
@@ -353,7 +410,7 @@
                   "<Add Type=\"" + typeof(StubTelemetryProcessor2).AssemblyQualifiedName + "\" />" +
                   "</TelemetryProcessors>");
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, new TestableTelemetryModules(), configFileContents);
 
             Assert.True(configuration.TelemetryProcessors != null);
@@ -375,7 +432,7 @@
                   </TelemetryProcessors>"
                 );
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null, configFileContents);
 
             // Assume that LoadFromXml method is called, tested separately
@@ -395,7 +452,7 @@
                   "<Add Type=\""+ typeof(StubTelemetryProcessor2).AssemblyQualifiedName + "\" />"+
                   "</TelemetryProcessors>");
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null, configFileContents);
 
             Assert.True(configuration.TelemetryProcessors != null);
@@ -415,7 +472,7 @@
                   </TelemetryProcessors>-->"
                 );
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null, configFileContents);
 
             // Assume that LoadFromXml method is called, tested separately
@@ -436,7 +493,7 @@
                   </TelemetryProcessors>"
                 );
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null, configFileContents);
 
             // Assume that LoadFromXml method is called, tested separately
@@ -454,7 +511,7 @@
                   </TelemetryProcessors>"
                 );
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null, configFileContents);
 
             configuration.TelemetryProcessorChainBuilder.Build();
@@ -473,7 +530,7 @@
                   </TelemetryProcessors>"
                 );
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null, configFileContents);
 
             var builder = configuration.TelemetryProcessorChainBuilder;
@@ -494,7 +551,7 @@
                   </TelemetryProcessors>"
                 );
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null, configFileContents);
 
             var builder = configuration.TelemetryProcessorChainBuilder;
@@ -595,7 +652,7 @@
                   </TelemetryInitializers>"
                 );
 
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null, configFileContents);
 
             Assert.Equal(2, configuration.TelemetryInitializers.Count); // Time and operation initializers are added by default
@@ -625,7 +682,7 @@
         [TestMethod]
         public void LoadInstancesUpdatesInstanceWithMatchingType()
         {
-            var configuration = new TelemetryConfiguration();
+            TelemetryConfiguration configuration = new TelemetryConfiguration();
             var element = XElement.Parse(@"
                 <List xmlns=""http://schemas.microsoft.com/ApplicationInsights/2013/Settings"">
                     <Add Type=""" + typeof(StubConfigurableWithProperties).AssemblyQualifiedName + @""" >