+++ Test/CoreSDK.Test/Shared/Extensibility/TelemetryConfigurationTest.cs
@@ -101,7 +101,7 @@
                 var tasks = new Task[8];
                 for (int i = 0; i < tasks.Length; i++)
                 {
-                    tasks[i] = TaskEx.Run(() => TelemetryConfiguration.Active);
+                    tasks[i] = Task.Run(() => TelemetryConfiguration.Active);
                 }
 
                 Task.WaitAll(tasks);