+++ src/TelemetryChannels/ServerTelemetryChannel/Shared/Implementation/TransmissionSender.cs
@@ -1,6 +1,10 @@
 ﻿namespace Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel.Implementation
 {
     using System;
+    using System.Collections.Generic;
+    using System.IO;
+    using System.Net;
+    using System.Runtime.Serialization;
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.ApplicationInsights.Channel;
@@ -10,6 +14,12 @@
     {
         private int transmissionCount = 0;
         private int capacity = 3;
+
+        private bool applyThrottle = false;
+        private int throttleWindowInMilliseconds = 1000;
+        private long currentThrottleWindowId = 0;
+        private int currentItemsCount = 0;
+        private int throttleLimit = 10;
 
         public event EventHandler<TransmissionProcessedEventArgs> TransmissionSent;
 
@@ -36,6 +46,56 @@
                 }
 
                 this.capacity = value;
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether a limiter on the maximum number of <see cref="ITelemetry"/> objects 
+        /// that can be sent in a given throttle window is enabled. Items attempted to be sent exceeding of the local 
+        /// throttle amount will be treated the same as a backend throttle.
+        /// </summary>
+        public virtual bool ApplyThrottle
+        {
+            get
+            {
+                return this.applyThrottle;
+            }
+
+            set
+            {
+                this.applyThrottle = value;
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets the maximum number of items that will be allowed to send in a given throttle window.
+        /// </summary>
+        public virtual int ThrottleLimit
+        {
+            get
+            {
+                return this.throttleLimit;
+            }
+
+            set
+            {
+                this.throttleLimit = value;
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets the size of the self-limiting throttle window in milliseconds.
+        /// </summary>
+        public virtual int ThrottleWindow
+        {
+            get
+            {
+                return this.throttleWindowInMilliseconds;
+            }
+
+            set
+            {
+                this.throttleWindowInMilliseconds = value;
             }
         }
 
@@ -86,10 +146,14 @@
             Exception exception = null;
             HttpWebResponseWrapper responseContent = null;
 
+            // Locally self-throttle this payload before we send it
+            Transmission acceptedTransmission = this.Throttle(transmission);
+
+            // Now that we've self-imposed a throttle, we can try to send the remaining data
             try
             {
-                TelemetryChannelEventSource.Log.TransmissionSendStarted(transmission.Id);
-                responseContent = await transmission.SendAsync().ConfigureAwait(false);          
+                TelemetryChannelEventSource.Log.TransmissionSendStarted(acceptedTransmission.Id);
+                responseContent = await acceptedTransmission.SendAsync().ConfigureAwait(false);          
             }
             catch (Exception e)
             {
@@ -100,15 +164,108 @@
                 int currentCapacity = Interlocked.Decrement(ref this.transmissionCount);
                 if (exception == null)
                 {
-                    TelemetryChannelEventSource.Log.TransmissionSentSuccessfully(transmission.Id, currentCapacity);
+                    TelemetryChannelEventSource.Log.TransmissionSentSuccessfully(acceptedTransmission.Id, currentCapacity);
                 }
                 else
                 {
-                    TelemetryChannelEventSource.Log.TransmissionSendingFailedWarning(transmission.Id, exception.ToString());
+                    TelemetryChannelEventSource.Log.TransmissionSendingFailedWarning(acceptedTransmission.Id, exception.ToString());
                 }
 
-                this.OnTransmissionSent(new TransmissionProcessedEventArgs(transmission, exception, responseContent));
+                if (responseContent == null && exception is WebException)
+                {
+                    HttpWebResponse response = (HttpWebResponse)((WebException)exception).Response;
+                    responseContent = new HttpWebResponseWrapper()
+                    {
+                        StatusCode = (int)response.StatusCode,
+                        StatusDescription = response.StatusDescription,
+                        RetryAfterHeader = response.Headers?.Get("Retry-After")
+                    };
+                }
+
+                this.OnTransmissionSent(new TransmissionProcessedEventArgs(acceptedTransmission, exception, responseContent));
             }
+        }
+
+        /// <summary>
+        /// Checks if the transmission throttling policy allows for sending another request.
+        /// If so, this method will add a request to the current throttle count (unless peeking).
+        /// </summary>
+        /// <returns>The number of events that are able to be sent.</returns>
+        private int IsTransmissionSendable(int numEvents, bool peek = false)
+        {
+            if (!this.applyThrottle)
+            {
+                return numEvents;
+            }
+
+            long throttleWindowId = (long)(DateTime.UtcNow.Subtract(new DateTime(1970, 1, 1)).TotalMilliseconds / this.ThrottleWindow);
+            bool isInThrottleWindow = throttleWindowId == this.currentThrottleWindowId;
+            if (isInThrottleWindow && this.currentItemsCount < this.ThrottleLimit)
+            {
+                int numAccepted = Math.Min(numEvents, this.ThrottleLimit - this.currentItemsCount);
+                if (!peek)
+                {
+                    this.currentItemsCount += numAccepted;
+                }
+
+                return numAccepted;
+            }
+            else if (!isInThrottleWindow)
+            {
+                int numAccepted = Math.Min(numEvents, this.ThrottleLimit);
+                this.currentThrottleWindowId = throttleWindowId;
+                this.currentItemsCount = numAccepted;
+                return numAccepted;
+            }
+
+            return 0;
+        }
+
+        private Transmission Throttle(Transmission transmission)
+        {
+            if (!this.ApplyThrottle)
+            {
+                return transmission;
+            }
+
+            int attemptedItemsCount = -1;
+            int acceptedItemsCount = -1;
+            Tuple<Transmission, Transmission> transmissions = transmission.Split((transmissionLength) => 
+            {
+                attemptedItemsCount = transmissionLength;
+                acceptedItemsCount = this.IsTransmissionSendable(transmissionLength);
+                return acceptedItemsCount;
+            });
+
+            Transmission acceptedTransmission = transmissions.Item1;
+            Transmission rejectedTransmission = transmissions.Item2;
+            
+            // Send rejected payload back for retry
+            if (rejectedTransmission != null)
+            {
+                TelemetryChannelEventSource.Log.TransmissionThrottledWarning(this.ThrottleLimit, attemptedItemsCount, acceptedItemsCount);
+                this.SendTransmissionThrottleRejection(rejectedTransmission);
+            }
+            
+            return acceptedTransmission;
+        }
+
+        private void SendTransmissionThrottleRejection(Transmission rejectedTransmission)
+        {
+            WebException exception = new WebException(
+                "Transmission was split by local throttling policy",
+                null,
+                System.Net.WebExceptionStatus.Success,
+                null);
+            this.OnTransmissionSent(new TransmissionProcessedEventArgs(
+                rejectedTransmission, 
+                exception, 
+                new HttpWebResponseWrapper()
+                {
+                    StatusCode = ResponseStatusCodes.ResponseCodeTooManyRequests,
+                    StatusDescription = "Internally Throttled",
+                    RetryAfterHeader = null
+                }));
         }
     }
 }